+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Add `ValueUtil.addSeconds(...)` utility method.
+* Add `ValueUtil.minus...()` utility methods for every `ValueUtil.add...()` method.
 
 ### [v0.105](https://github.com/realityforge/guiceyloops/tree/v0.105) (2019-09-25) · [Full Changelog](https://github.com/realityforge/guiceyloops/compare/v0.104...v0.105)